refactor : made components
