+++ src/components/MainFocusToday.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const MainFocusToday = () => {
+  return (
+    <p className="text-3xl font-black text-white capitalize mt-3">
+      what is your main focus today?
+    </p>
+  );
+}
