+++ src/components/index.js
@@ -1,0 +1,6 @@
+export {TimeGreet} from '../components/TimeGreet';
+export {Weather} from '../components/Weather';
+export {MainTask} from '../components/MainTask';
+export {EditDeleteMainTask} from '../components/EditDeleteMainTask';
+export {Quote} from '../components/Quote';
+export {MainFocusToday} from '../components/MainFocusToday';
